requirements.txt
nicegui
google-cloud-storage